Create the "Tracks" DATABASE
I wrote a code that connect with the local database and executes SQL code with cur.executescript() function.
First of all, the code delete the table if exists. This way allows me to create the table from scratch once and again.

The, I create three tables:

1.Artist table: This table will keep track of every singer or musical group that is in the XML file that will be parsed.

2.Album table: This will keep track of every track album that is in the XML file. This table also has an "artist_id"
attribute to make a connection between the artist table and the album table.

3.Tracks table: This will keep track of every track name with its respective len, rating, count and it also has an
"album_id" as foreign attribute to make a connection with the Album table.