Parsing the XML tracks file
First, I parsed the xml file into an object stored in the tree variable. Then, I found all the dict tags that contain
the important information about the track with the findall method available in the tree object. This dict tag is a
'three order' dict, so, its tag path is ./dict/dict/dict.

The lookup function is responsible of returning the tag value that is next to the key tag (in the xml file, these tags
are <string> and <integer>. If the function doesn't find the key tag value needed, this return None.

Finally, I iterate throgh every dictionary tag and use the lookup function in every dict. If there is no track or no
artist, album or track name, I'm not interested in that dictionary