Adding all the values to the database
First, I had a mistake and I made a new table called Genre and added a new column to tracks called genre_id. I also
had to use the lookup function to return the tag value into the genre variable.

With cur.execute I run SQL code that insert the content of the artist variable into the Artist table or ignore it if
the value is already in the table. I also select the id value of each artist and store it in artist_id variable, so I
can add this artist_id into the Album Table. This workflow is the same for each table.