Finishing the code: Parsing JSON and adding a Many-to-Many Relatinship to the Database
A tinny change was renaming the table at line 10, I had 'Member' But I updated this to Members. This table is used to
connect the to tables in a many-to-many relationship.

json.loads(string) returns a list that I iterated through a for loop and then adding each element of a nested list contained
in iteration variable 'row' in its respective variable (name, course, role). Then I executed some SQL with cur.execute
and stored the id primare key given by the database, and in this way I can use it later when I want to add it to 'Members'
table.